Overrides function parseDaiaArray.
Overrides function parseDaiaArray of DAIA class in PICA class. Possibly a better way would be to place this function in Hebis driver class or even better to build a new DAIA driver as well as to remove the PICA driver completly and move the whole functionality of that to the Hebis driver (cause PICA driver it's not longer a part of VuFind).